Follow de follow & trie par score Affiche les codes des follow de follow chainé. Trie par score ajoutant un champ score au post, incrémentation faite avec un lock optimiste.

+++ src/main/java/fr/uge/revevue/controller/rest/CodeRestController.java
@@ -26,7 +26,7 @@ public class CodeRestController {
     }
 
     @PostMapping("/vote/{codeId}")
-    public ResponseEntity<Integer> codeVoted(@PathVariable("codeId") @Valid long codeId,
+    public ResponseEntity<Long> codeVoted(@PathVariable("codeId") @Valid long codeId,
                                              @RequestParam("voteType") Vote.VoteType voteType) {
         return ResponseEntity.ok(voteService.postVoted(userService.currentUser().getId(), codeId, voteType));
     }
